# services/loop-plane/openapi.yaml
openapi: 3.1.0
info:
  title: loop-plane
  version: '1.0'
  description: Loop Plane 微服务 — 长程状态/因果/硬化干预/审计/调度门面 (SPEC v1.0)
paths:
...
  /v1/loop/skills/{skill_id}/experiments:
    post:
      tags:
      - skills
      summary: Start Experiment
      operationId: start_experiment_v1_loop_skills__skill_id__experiments_post
      parameters:
      - name: skill_id
        in: path
        required: true
        schema:
          type: string
          title: Skill Id
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Start Experiment V1 Loop Skills  Skill Id  Experiments
                  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/loop/skills/experiments/{experiment_id}/optimize:
    post:
      tags:
      - skills
      summary: Optimize Experiment
      operationId: optimize_experiment_v1_loop_skills_experiments__experiment_id__optimize_post
      parameters:
      - name: experiment_id
        in: path
        required: true
        schema:
          type: string
          title: Experiment Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Optimize Experiment V1 Loop Skills Experiments  Experiment
                  Id  Optimize Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/loop/skills/{skill_id}/release:
    post:
      tags:
      - skills
      summary: Release Skill
      operationId: release_skill_v1_loop_skills__skill_id__release_post
      parameters:
      - name: skill_id
        in: path
        required: true
        schema:
          type: string
          title: Skill Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Release Skill V1 Loop Skills  Skill Id  Release Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/loop/skills/{skill_id}/rollback:
    post:
      tags:
      - skills
      summary: Rollback Skill
      operationId: rollback_skill_v1_loop_skills__skill_id__rollback_post
      parameters:
      - name: skill_id
        in: path
        required: true
        schema:
          type: string
          title: Skill Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Rollback Skill V1 Loop Skills  Skill Id  Rollback
                  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError